# fix: Mobile Performance Optimization v2 (#359)