--- a/phialo-design/src/features/portfolio/components/OptimizedPortfolioGrid.tsx
+++ b/phialo-design/src/features/portfolio/components/OptimizedPortfolioGrid.tsx
@@ -1,0 +1,221 @@
+import React, { useCallback, useEffect, useState } from 'react';
+import { motion, AnimatePresence } from '../../../lib/framer-motion';
+import MagneticCursor from '../../../shared/components/effects/MagneticCursor';
+import type { PortfolioItemData } from './PortfolioSection';
+
+interface OptimizedPortfolioGridProps {
+  items: PortfolioItemData[];
+  onItemClick?: (item: PortfolioItemData) => void;
+  lang?: 'en' | 'de';
+}
+
+// Grid container variants for staggered animation
+const gridVariants = {
+  hidden: { opacity: 0 },
+  visible: {
+    opacity: 1,
+    transition: {
+      staggerChildren: 0.05, // Reduced stagger for faster perceived loading
+      delayChildren: 0.1,
+    },
+  },
+};
+
+// Item variants with reduced motion complexity
+const itemVariants = {
+  hidden: { 
+    opacity: 0, 
+    y: 20, // Reduced distance for smoother animation
+  },
+  visible: {
+    opacity: 1,
+    y: 0,
+    transition: {
+      type: "spring" as const,
+      damping: 25, // Increased damping for less bouncy animation
+      stiffness: 120, // Slightly increased for snappier response
+      mass: 0.5, // Reduced mass for lighter feel
+    },
+  },
+};
+
+function OptimizedPortfolioItem({ 
+  item, 
+  index, 
+  onItemClick, 
+  lang = 'de' 
+}: { 
+  item: PortfolioItemData; 
+  index: number;
+  onItemClick?: (item: PortfolioItemData) => void; 
+  lang?: 'en' | 'de';
+}) {
+  const isEnglish = lang === 'en';
+  const [imageLoaded, setImageLoaded] = useState(false);
+  const [imageError, setImageError] = useState(false);
+
+  // Determine if this is a priority image
+  const isPriority = index < 3;
+
+  // Keyboard handler for accessibility
+  const handleKeyDown = useCallback((event: React.KeyboardEvent) => {
+    if (event.key === 'Enter' || event.key === ' ') {
+      event.preventDefault();
+      onItemClick?.(item);
+    }
+  }, [item, onItemClick]);
+
+  // Handle image load events
+  const handleImageLoad = useCallback(() => {
+    setImageLoaded(true);
+  }, []);
+
+  const handleImageError = useCallback(() => {
+    setImageError(true);
+    console.error(`Failed to load image: ${item.image}`);
+  }, [item.image]);
+
+  // Check if image is already loaded (from cache)
+  useEffect(() => {
+    const img = new Image();
+    img.src = item.image;
+    
+    if (img.complete && img.naturalWidth > 0) {
+      setImageLoaded(true);
+    } else {
+      img.onload = handleImageLoad;
+      img.onerror = handleImageError;
+    }
+
+    return () => {
+      img.onload = null;
+      img.onerror = null;
+    };
+  }, [item.image, handleImageLoad, handleImageError]);
+
+  // Translations for accessibility
+  const cardLabel = isEnglish 
+    ? `Portfolio item: ${item.title}. Category: ${item.category}. Press Enter or Space to view details.`
+    : `Portfolio-Element: ${item.title}. Kategorie: ${item.category}. Drücken Sie Enter oder Leertaste, um Details anzuzeigen.`;
+
+  return (
+    <motion.div
+      variants={itemVariants}
+      layout
+      className="portfolio-item"
+      style={{
+        // CSS containment for performance
+        contain: 'layout style paint',
+      }}
+    >
+      <div
+        className="group relative cursor-pointer focus:outline-none focus:ring-2 focus:ring-gold focus:ring-offset-2 rounded-lg"
+        data-testid="portfolio-item"
+        data-item-id={item.id}
+        role="button"
+        tabIndex={0}
+        aria-label={cardLabel}
+        onClick={() => onItemClick?.(item)}
+        onKeyDown={handleKeyDown}
+      >
+        <MagneticCursor>
+          <div className="portfolio-item-container overflow-hidden rounded-lg bg-gray-100 aspect-[4/5] transition-all duration-300 group-hover:shadow-lg">
+            {/* Show placeholder while image loads */}
+            {!imageLoaded && !imageError && (
+              <div className="absolute inset-0 bg-gradient-to-br from-gray-200 to-gray-100 animate-pulse" />
+            )}
+            
+            {/* Error state */}
+            {imageError && (
+              <div className="absolute inset-0 flex items-center justify-center bg-gray-100">
+                <span className="text-gray-400">Failed to load image</span>
+              </div>
+            )}
+            
+            {/* Optimized image */}
+            <img
+              src={item.image}
+              alt={`${item.title} - ${item.category}`}
+              className={`w-full h-full object-cover transition-all duration-300 group-hover:scale-105 group-focus:scale-105 ${
+                imageLoaded ? 'opacity-100' : 'opacity-0'
+              }`}
+              loading={isPriority ? 'eager' : 'lazy'}
+              decoding={isPriority ? 'sync' : 'async'}
+              fetchPriority={isPriority ? 'high' : undefined}
+              onLoad={handleImageLoad}
+              onError={handleImageError}
+              width="800"
+              height="1000"
+            />
+
+            {/* Overlay */}
+            <div className="absolute inset-0 bg-gradient-to-t from-midnight/80 via-midnight/20 to-transparent opacity-0 group-hover:opacity-100 group-focus:opacity-100 transition-opacity duration-300 pointer-events-none">
+              <div className="absolute bottom-0 left-0 right-0 p-6 text-white">
+                <div className="bg-black/50 rounded-lg p-4 backdrop-blur-sm">
+                  <h3 className="font-display text-3xl font-medium mb-2">{item.title}</h3>
+                  <p className="text-lg text-white/90">{item.category}</p>
+                </div>
+              </div>
+            </div>
+          </div>
+        </MagneticCursor>
+      </div>
+    </motion.div>
+  );
+}
+
+export default function OptimizedPortfolioGrid({ 
+  items, 
+  onItemClick, 
+  lang = 'de' 
+}: OptimizedPortfolioGridProps) {
+  // Check for reduced motion preference
+  const [prefersReducedMotion, setPrefersReducedMotion] = useState(false);
+
+  useEffect(() => {
+    const mediaQuery = window.matchMedia('(prefers-reduced-motion: reduce)');
+    setPrefersReducedMotion(mediaQuery.matches);
+
+    const handleChange = (event: MediaQueryListEvent) => {
+      setPrefersReducedMotion(event.matches);
+    };
+
+    mediaQuery.addEventListener('change', handleChange);
+    return () => mediaQuery.removeEventListener('change', handleChange);
+  }, []);
+
+  // Use simpler animations if reduced motion is preferred
+  const effectiveVariants = prefersReducedMotion
+    ? { hidden: { opacity: 0 }, visible: { opacity: 1 } }
+    : gridVariants;
+
+  const _effectiveItemVariants = prefersReducedMotion
+    ? { hidden: { opacity: 0 }, visible: { opacity: 1 } }
+    : itemVariants;
+
+  return (
+    <motion.div
+      className="portfolio-grid"
+      data-testid="portfolio-grid"
+      variants={effectiveVariants}
+      initial="hidden"
+      animate="visible"
+      style={{
+        // Prevent layout shifts
+        minHeight: items.length > 0 ? '500px' : 'auto',
+      }}
+    >
+      <AnimatePresence mode="popLayout">
+        {items.map((item, index) => (
+          <OptimizedPortfolioItem
+            key={item.id}
+            item={item}
+            index={index}
+            onItemClick={onItemClick}
+            lang={lang}
+          />
+        ))}
+      </AnimatePresence>
+    </motion.div>
+  );
+}