--- a/phialo-design/src/pages/test-mobile-optimization.astro
+++ b/phialo-design/src/pages/test-mobile-optimization.astro
@@ -1,0 +1,180 @@
+---
+import MobileOptimizedLayout from '../shared/layouts/MobileOptimizedLayout.astro';
+import OptimizedPortfolioGrid from '../features/portfolio/components/OptimizedPortfolioGrid';
+import { portfolioItemsDE } from '../features/portfolio/data/portfolioDataDE';
+import { portfolioItemsEN } from '../features/portfolio/data/portfolioDataEN';
+
+// Get language from URL
+const lang = Astro.url.pathname.startsWith('/en/') ? 'en' : 'de';
+const portfolioItems = lang === 'en' ? portfolioItemsEN : portfolioItemsDE;
+
+// Get first image for LCP preload
+const priorityImageSrc = portfolioItems[0]?.image;
+
+// Get first 3 images for additional preloading
+const preloadImages = portfolioItems.slice(0, 3).map(item => item.image);
+
+const title = lang === 'en' 
+  ? 'Mobile Optimization Test - Phialo Design' 
+  : 'Mobile-Optimierung Test - Phialo Design';
+
+const description = lang === 'en'
+  ? 'Testing mobile performance optimizations for the portfolio gallery'
+  : 'Test der Mobile-Performance-Optimierungen für die Portfolio-Galerie';
+---
+
+<MobileOptimizedLayout 
+  title={title} 
+  description={description}
+  lang={lang}
+  priorityImageSrc={priorityImageSrc}
+  preloadImages={preloadImages}
+>
+  <main class="min-h-screen bg-white">
+    <!-- Header -->
+    <header class="bg-midnight text-white py-8 px-4">
+      <div class="container mx-auto">
+        <h1 class="font-display text-4xl md:text-5xl lg:text-6xl font-bold mb-4">
+          {lang === 'en' ? 'Mobile Optimization Test' : 'Mobile-Optimierung Test'}
+        </h1>
+        <p class="text-lg text-white/80">
+          {lang === 'en' 
+            ? 'Testing performance improvements for mobile devices' 
+            : 'Test der Leistungsverbesserungen für mobile Geräte'}
+        </p>
+      </div>
+    </header>
+
+    <!-- Performance Metrics Info -->
+    <section class="bg-gray-50 py-6 px-4">
+      <div class="container mx-auto">
+        <div class="bg-white rounded-lg shadow-sm p-6">
+          <h2 class="font-display text-2xl font-semibold mb-4 text-midnight">
+            {lang === 'en' ? 'Optimizations Applied' : 'Angewendete Optimierungen'}
+          </h2>
+          <ul class="space-y-2 text-gray-700">
+            <li class="flex items-start">
+              <span class="text-green-500 mr-2">✓</span>
+              <span>Priority loading for first 3 images (LCP optimization)</span>
+            </li>
+            <li class="flex items-start">
+              <span class="text-green-500 mr-2">✓</span>
+              <span>CSS containment for rendering isolation</span>
+            </li>
+            <li class="flex items-start">
+              <span class="text-green-500 mr-2">✓</span>
+              <span>Safe image loading with cache detection</span>
+            </li>
+            <li class="flex items-start">
+              <span class="text-green-500 mr-2">✓</span>
+              <span>Reduced animation complexity on mobile</span>
+            </li>
+            <li class="flex items-start">
+              <span class="text-green-500 mr-2">✓</span>
+              <span>Touch-optimized interactions (44x44px targets)</span>
+            </li>
+            <li class="flex items-start">
+              <span class="text-green-500 mr-2">✓</span>
+              <span>Resource hints and preloading</span>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </section>
+
+    <!-- Optimized Portfolio Grid -->
+    <section class="py-12 px-4">
+      <div class="container mx-auto">
+        <h2 class="font-display text-3xl md:text-4xl font-bold mb-8 text-midnight text-center">
+          {lang === 'en' ? 'Optimized Portfolio Gallery' : 'Optimierte Portfolio-Galerie'}
+        </h2>
+        
+        <!-- The optimized grid component -->
+        <OptimizedPortfolioGrid 
+          items={portfolioItems}
+          lang={lang}
+          client:visible
+          onItemClick={(item) => {
+            console.log('Portfolio item clicked:', item);
+            // Modal functionality would be added here
+          }}
+        />
+      </div>
+    </section>
+
+    <!-- Testing Instructions -->
+    <section class="bg-gray-50 py-12 px-4">
+      <div class="container mx-auto">
+        <div class="bg-white rounded-lg shadow-sm p-6">
+          <h2 class="font-display text-2xl font-semibold mb-4 text-midnight">
+            {lang === 'en' ? 'Testing Instructions' : 'Testanweisungen'}
+          </h2>
+          <ol class="space-y-3 text-gray-700">
+            <li>
+              <strong>1. Open Chrome DevTools</strong>
+              <p class="text-sm text-gray-600 mt-1">Press F12 or right-click → Inspect</p>
+            </li>
+            <li>
+              <strong>2. Switch to Mobile View</strong>
+              <p class="text-sm text-gray-600 mt-1">Click device toolbar icon or press Ctrl+Shift+M</p>
+            </li>
+            <li>
+              <strong>3. Run Lighthouse Audit</strong>
+              <p class="text-sm text-gray-600 mt-1">Go to Lighthouse tab → Select "Mobile" → Generate report</p>
+            </li>
+            <li>
+              <strong>4. Enable Network Throttling</strong>
+              <p class="text-sm text-gray-600 mt-1">Network tab → Select "Slow 3G" to test on slow connections</p>
+            </li>
+            <li>
+              <strong>5. Check Console for Errors</strong>
+              <p class="text-sm text-gray-600 mt-1">Console tab → Look for hydration warnings or errors</p>
+            </li>
+          </ol>
+          
+          <div class="mt-6 p-4 bg-blue-50 rounded-lg">
+            <p class="text-sm text-blue-800">
+              <strong>Expected Results:</strong> Lighthouse mobile score of 85+, 
+              no hydration errors, smooth scrolling, instant image visibility.
+            </p>
+          </div>
+        </div>
+      </div>
+    </section>
+
+    <!-- Comparison Link -->
+    <section class="py-8 px-4 text-center">
+      <div class="container mx-auto">
+        <p class="text-gray-600 mb-4">
+          {lang === 'en' ? 'Compare with original:' : 'Vergleichen Sie mit dem Original:'}
+        </p>
+        <a 
+          href={lang === 'en' ? '/en/portfolio' : '/portfolio'}
+          class="inline-block bg-midnight text-white px-6 py-3 rounded-lg hover:bg-midnight/90 transition-colors"
+        >
+          {lang === 'en' ? 'View Original Portfolio' : 'Original-Portfolio anzeigen'}
+        </a>
+      </div>
+    </section>
+  </main>
+</MobileOptimizedLayout>
+
+<style>
+  /* Page-specific styles */
+  .container {
+    max-width: 1200px;
+  }
+  
+  /* Ensure smooth scrolling */
+  html {
+    scroll-behavior: smooth;
+  }
+  
+  /* Performance monitoring helper */
+  @media (max-width: 768px) {
+    .portfolio-grid {
+      /* Single column on mobile */
+      grid-template-columns: 1fr !important;
+    }
+  }
+</style>