… phialo-design/docs/mobile-optimization-guide.md
… phialo-design/src/shared/layouts/MobileOptimizedLayout.astro
… phialo-design/src/styles/mobile-optimizations.css