@barde 2025-08-16
## 🚀 Mobile Performance Optimizations Update

I've implemented comprehensive mobile performance optimizations based on the Lighthouse report analysis. Here are the key improvements:

### ✅ Implemented Optimizations

#### 1. **Build-Time Image Optimization** 🖼️
- Moved portfolio images to `src/assets/images/portfolio/` for Astro's build-time processing
- Created `OptimizedImage.astro` component that leverages Astro's built-in Image component
- This approach automatically generates WebP/AVIF formats during build, eliminating the need for manual conversion
- Images are only checked into Git LFS in their original format, with optimized versions generated at build time

#### 2. **Adaptive Loading System** 🌐
- Implemented `AdaptiveLoader` utility that detects:
  - Network speed (slow/medium/fast)
  - Device capability (low/medium/high)
  - User preferences (reduced motion, data saving)
- Automatically adjusts:
  - Animation complexity and duration
  - Number of images to preload
  - Resource loading strategies

#### 3. **Mobile Touch & Performance Optimizations** 📱
- Enhanced touch target sizes (minimum 44x44px)
- Improved scrolling performance with hardware acceleration
- Added proper CSS containment for better rendering performance
- Disabled hover effects on touch devices to prevent sticky states
- Implemented reduced motion support

#### 4. **SSR-Safe Implementation** 🛡️
- All adaptive loading hooks are SSR-safe with proper fallbacks
- Window checks prevent build-time errors
- Graceful degradation for server-side rendering

### 📊 Expected Performance Improvements

Based on the optimizations:
- **Image Size Reduction**: ~60-80% with WebP/AVIF formats
- **First Contentful Paint**: ~40% improvement
- **Largest Contentful Paint**: ~50% improvement
- **Mobile Lighthouse Score**: Target 85-95 (from current ~74)

### 🔄 CI Status
All tests are passing ✅

### 📝 Note on Implementation
As suggested, I've switched from manual image conversion to Astro's build-time optimization. This means:
- Original images stay in Git LFS
- Optimized formats are generated during build
- No need to maintain multiple image formats in the repository
- Automatic srcset and responsive image generation

The preview deployment should now show improved mobile performance. You can test it at: https://phialo-pr-366.meise.workers.dev

Ready for review and Lighthouse re-testing! 🎯