@barde 2025-08-16
Already better! Please take care of the remaining issues according to Lighthouse.

[phialo-pr-366.meise.workers.dev-20250816T125720.json](https://github.com/user-attachments/files/21811838/phialo-pr-366.meise.workers.dev-20250816T125720.json)