@barde 2025-08-16
Still issues, especially "Improve image delivery" and "Network dependency tree"!

[phialo-pr-366.meise.workers.dev-20250816T132254.json](https://github.com/user-attachments/files/21812213/phialo-pr-366.meise.workers.dev-20250816T132254.json)