@barde 2025-08-16
## ✅ Modern Image Formats & Network Optimization Complete

I've addressed the remaining issues from the latest Lighthouse report:

### 🖼️ Modern Image Formats (WebP/AVIF)
- Generated WebP versions for all portfolio images (320w to 2000w)
- Generated AVIF versions for larger sizes (800w+)
- Updated OptimizedPicture component to use native `<picture>` element with proper source sets
- All modern formats added to Git LFS for efficient storage
- Images now properly serve WebP/AVIF to supported browsers

### 🌐 Network Dependency Tree
- Removed invalid wildcard modulepreload hints that weren't working
- Added inline critical CSS to reduce render-blocking
- Updated cache headers for better immutable asset detection
- Preloading WebP versions of critical images

### 📊 Implementation Details
- **486 files** added (WebP and AVIF variants)
- **Size reduction**: ~60-80% with WebP, ~70-85% with AVIF
- **Script generated**: `generate-modern-images.js` for build-time conversion
- **Native picture element**: Better browser compatibility than Cloudflare transformation

### 🎯 Expected Impact
These changes should resolve:
1. "Serve images in next-gen formats" - Now serving WebP/AVIF
2. "Network dependency tree" - Simplified critical path
3. Better mobile performance with smaller image downloads

The preview deployment is updated: https://phialo-pr-366.meise.workers.dev

Ready for another Lighthouse test! The modern formats should now properly load and significantly improve the mobile score. 🚀