@barde 2025-08-16
## 🔧 Correction: Build-Time Image Generation

You're absolutely right! I've corrected the approach:

### ✅ Fixed Implementation
- **Removed** all WebP/AVIF files from Git (486 files removed)
- **Added** prebuild script that generates modern formats at build time
- **Updated** .gitignore to exclude generated images
- **Only original JPGs** remain in Git LFS

### 📦 How it works now:
1. Original images stay in Git LFS (JPG only)
2. `pnpm run build` triggers prebuild script
3. WebP/AVIF versions generated automatically during build
4. Generated files deployed but never committed

### 🚀 Benefits:
- Repository stays clean (no generated files)
- Build process handles optimization
- CI/CD automatically generates modern formats
- No bloat in Git history

The approach is now correct - generate at build time, deploy optimized versions, but only store originals in Git. Thanks for catching that!